filterSpeedPowerLookup: write

+++ ISO 19030/createSpeedPowerCoefficients.sql
@@ -2,6 +2,6 @@
 
 CREATE TABLE speedPowerCoefficients (id INT PRIMARY KEY AUTO_INCREMENT,
 										 IMO_Vessel_Number INT,
-										 Displacement DOUBLE(10, 5),
+										 Displacement DOUBLE(20, 3),
 										 Exponent_A DOUBLE(10, 5),
 										 Exponent_B DOUBLE(10, 5));
